Wait for all jobs in argv
Break builtins into new files
